data-extraction/pyproject.toml
[tool.poetry]
name = "statsfin-data-extractor"
version = "0.1.0"
description = "This Python package extracts data from StatsFin API for postal code area analysis."
authors = ["shiftleino <shiftleino@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
requests = "^2.31.0"
pandas = "^2.1.1"
numpy = "^1.26.1"
google-cloud-storage = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
